requirements.txt
websockets
requests
protobuf
python-dotenv
pymongo
